+++ MonsterHunterMobile/Assets/Scripts/Socket/Servers/Controller/GameController.cs
@@ -1,0 +1,30 @@
+﻿using SocketGameProtocol;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+class GameController : BaseController
+{
+
+    public GameController()
+    {
+        requestCode = RequestCode.Game;
+    }
+
+    //更新玩家位置
+    public MainPack UpPos(Client client, MainPack pack)
+    {
+        client.GetRoom.BroadcastTo(client, pack);
+        client.UpPos(pack);//更新位置信息
+        return null;
+    }
+
+    //更新玩家操作
+    public MainPack UpOperation(Client client, MainPack pack)
+    {
+
+        return null;
+    }
+}
