+++ MonsterHunterMobile/Assets/Scripts/Socket/UIPanel/StartPanel.cs
@@ -20,6 +20,7 @@ public class StartPanel : BasePanel
     private void StartButtonClick()
     {
         UIManager.Instance.PushPanel(PanelName.RoomList,PanelType.normalPanel);
+        //RequestCenter.GetInstance().CreateRoom(roomName.text, num);
     }
 
     private void ExitButtonClick()
