+++ MonsterHunterMobile/GlobalSuppressions.cs
@@ -1,0 +1,18 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Build", "UNT0001:The Unity message \"Start\" is empty.", Justification = "<挂起>", Scope = "member", Target = "~M:DemoKeyCommands.Start")]
+[assembly: SuppressMessage("Build", "UNT0002:Comparing tags using == is inefficient.", Justification = "<挂起>", Scope = "member", Target = "~M:DM.ControlManager.OnTriggerExit(UnityEngine.Collider)")]
+[assembly: SuppressMessage("Build", "UNT0002:Comparing tags using == is inefficient.", Justification = "<挂起>", Scope = "member", Target = "~M:DM.ControlManager.OnTriggerStay(UnityEngine.Collider)")]
+[assembly: SuppressMessage("Build", "UNT0001:The Unity message \"Start\" is empty.", Justification = "<挂起>", Scope = "member", Target = "~M:HeroInfoUI.Start")]
+[assembly: SuppressMessage("Build", "UNT0001:The Unity message \"Update\" is empty.", Justification = "<挂起>", Scope = "member", Target = "~M:HeroInfoUI.Update")]
+[assembly: SuppressMessage("Build", "UNT0001:The Unity message \"Start\" is empty.", Justification = "<挂起>", Scope = "member", Target = "~M:Loom.Start")]
+[assembly: SuppressMessage("Build", "UNT0001:The Unity message \"Update\" is empty.", Justification = "<挂起>", Scope = "member", Target = "~M:PlayerBox.Update")]
+[assembly: SuppressMessage("Build", "UNT0001:The Unity message \"Start\" is empty.", Justification = "<挂起>", Scope = "member", Target = "~M:PlayerMessage.Start")]
+[assembly: SuppressMessage("Build", "UNT0001:The Unity message \"Update\" is empty.", Justification = "<挂起>", Scope = "member", Target = "~M:PlayerMessage.Update")]
+[assembly: SuppressMessage("Build", "UNT0001:The Unity message \"Start\" is empty.", Justification = "<挂起>", Scope = "member", Target = "~M:SFB_DemoControlKeys.Start")]
+[assembly: SuppressMessage("Build", "UNT0013:SerializeField attribute is invalid or redundant for property or field: UserName", Justification = "<挂起>", Scope = "member", Target = "~P:ClientManager.UserName")]
